+++ src/index.ts
@@ -3,4 +3,8 @@ export { default } from "./lib/remarkWikiLink";
 export { syntax } from "./lib/syntax";
 export { fromMarkdown } from "./lib/fromMarkdown";
 export { toMarkdown } from "./lib/toMarkdown";
-export { defaultUrlResolver, findMatchingFilePath } from "./utils/index";
+export {
+  defaultUrlResolver,
+  defaultHeadingResolver,
+  findMatchingFilePath,
+} from "./utils/index";