+++ src/lib/remarkWikiLink.ts
@@ -15,6 +15,7 @@ export interface Options {
   className?: string; // class to be added to all wikilinks (and embeds)
   newClassName?: string; // class to added to wikilink (and embeds) that don't have matching files
   aliasDivider?: string; // (default: "|") character used to separate the target from the alias in wiki links during parsing and stringification
+  headingResolver?: (heading?: string) => string; // (default: (heading?: string) => heading ? `#${slug(heading)` : "") resolve headings when permalink is available
   urlResolver?: (opts: {
     filePath: string;
     isEmbed: boolean;